requirements.txt
PyAutoGUI==0.9.52
pytesseract==0.3.8
numpy==1.21.0
opencv_contrib_python==4.5.2.54
Pillow==8.3.1
